refactor: split browser endpoint to multiple files
